add user authentification & authorization with roles

+++ app/controllers/message.controller.js
@@ -4,7 +4,6 @@ const Message = db.messages;
 
 // POST / CREATE a new Message
 exports.create = (req, res) => { 
-  // const userId = req.params.userId;
   Message.create({  
     content: req.body.content,
     userId: req.body.userId,
@@ -46,27 +45,50 @@ exports.findByPk = (req, res) => {
 // Update a Message by Id (PUT / UPDATE)
 exports.update = (req, res) => {
   const id = req.params.messageId;
-  Message.update( { content: req.body.content }, 
-    { where: {id: req.params.messageId} }
-    ).then(() => {
-      res.status(200).send({ message: 'updated successfully a message with id = ' + id });
-    })
-    .catch((err) => {
-      console.log(">> Error while updating message with id: ", req.params.id, err);
-      res.send(err);
-    });
-  };
-  
-  // Delete a Message by Id (DELETE / DELETE)
-  exports.delete = (req, res) => {
-    const id = req.params.messageId;
-    Message.destroy({
-      where: { id: id }
-    }).then(() => {
-      res.status(200).send({ message: 'deleted successfully a message with id = ' + id });
-    })
-    .catch((err) => {
-      console.log(">> Error while deleting message with id: ", id, err);
-      res.send(err);
-    });
-  };
+  const currentUser = req.userId
+
+  Message.findOne({ id: id})
+  .then( (message) => {
+    if (message.userId === currentUser) {
+      Message.update( { content: req.body.content }, 
+        { where: { id: id }},
+      ).then( () => {
+        res.status(200).send({ message: 'updated successfully a message with id = ' + id + "user: " + req.userId });
+      })
+      .catch((err) => {
+        res.status(401).send(err);
+      });
+    } else {
+      res.status(401).send("You're not this message's author !");
+    }
+  })
+  .catch((err) => {
+    res.status(401).send(err);
+  });
+};
+
+// Delete a Message by Id (DELETE / DELETE)
+exports.delete = (req, res) => {
+  const id = req.params.messageId;
+  const currentUser = req.userId
+
+  Message.findOne({ id: id})
+  .then( (message) => {
+    if (message.userId === currentUser) {
+      Message.destroy({
+        where: { id: id }
+      }).then(() => {
+        res.status(200).send({ message: 'deleted successfully a message with id = ' + id });
+      })
+      .catch((err) => {
+        console.log(">> Error while deleting message with id: ", id, err);
+        res.send(err);
+      });
+    } else {
+      res.status(401).send("You're not this message's author !");
+    }
+  })
+  .catch((err) => {
+    res.status(401).send(err);
+  });
+};
